Rename new-compile-transaction
- function is already just `compileTransaction`, so just a file rename
