+++ packages/transactions/src/index.ts
@@ -1,6 +1,6 @@
 export * from './codecs';
 export * from './lifetime';
-export * from './new-compile-transaction';
+export * from './compile-transaction';
 export * from './new-signatures';
 export * from './new-wire-transaction';
 export * from './transaction';
